+++ src/daisy_versio.cpp
@@ -54,7 +54,7 @@ void DaisyVersio::Init(bool boost)
                          PIN_ADC_CV6};
 
     // gate in and momentary switch
-    tap.Init(seed.GetPin(PIN_SW), blockrate_);
+    tap.Init(seed.GetPin(PIN_SW));
     dsy_gpio_pin gate_gpio = seed.GetPin(PIN_TRIG_IN);
     gate.Init(&gate_gpio);
 
@@ -94,7 +94,6 @@ void DaisyVersio::SetHidUpdateRates()
     {
         knobs[i].SetSampleRate(AudioCallbackRate());
     }
-    tap.SetUpdateRate(AudioCallbackRate());
 }